requirements/backend.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pyspark==3.5.3
psycopg2-binary==2.9.9
redis==5.0.8
